requirements.txt
fastapi
uvicorn
requests
python-dotenv
elevenlabs
python-multipart
